+++ modules/card-project/card-project.php
@@ -7,12 +7,10 @@
 	));
 
 	$defaults = [
-		'tag' => 'div',
 		'permalink' => bp_get_group_permalink(),
 		'cover_image_url' => $cover_image_url,
 		'tela' => bp_groups_has_group_type( bp_get_group_id(), 'tela-botanica' ),
 		'tela_title' => __( 'Un projet Tela Botanica', 'telabotanica' ),
-		'tela_icon' => get_telabotanica_module('icon', ['icon' => 'tela-leaf']),
 		'avatar' => bp_core_fetch_avatar( [
 			'item_id' => bp_get_group_id(),
 			'object' => 'group',
@@ -34,7 +32,6 @@
 	$data->modifiers = telabotanica_styleguide_modifiers_array('card-project', $data->modifiers);
 
 	echo $pug->render(__DIR__ . '/card-project.pug', [
-		'tag' => $data->tag,
 		'data' => $data
 	]);