+++ modules/cover-search/cover-search.php
@@ -4,6 +4,8 @@
 	$defaults = [
 		'image' => get_field('cover_image'),
 		'total_results' => $wp_query->found_posts,
+		'index' => false,
+		'instantsearch' => false,
 		'modifiers' => []
 	];
 
@@ -21,6 +23,8 @@
 
 			the_telabotanica_module('search-box', [
 				'autocomplete' => false,
+				'instantsearch' => $data->instantsearch,
+				'index' => $data->index,
 				'modifiers' => ['large', 'is-primary']
 			]);
 
@@ -39,28 +43,7 @@
 			echo '</div>';
 		echo '</div>';
 
-		if ( $data->image ) :
-			$credits = get_fields( $data->image['ID'] );
-			if ( $credits ) :
-				echo '<div class="cover-credits">';
-					$title = $data->image['title'];
-
-					if ($credits['link']) {
-						$title = '<a href="' . $credits['link'] . '" target="_blank">' . $title . '</a>';
-					}
-
-					if ($credits['author']) {
-						printf(
-							__('%s par %s', 'telabotanica'),
-							$title,
-							$credits['author']
-						);
-					} else {
-						echo $title;
-					}
-				echo '</div>';
-			endif;
-		endif;
+		telabotanica_image_credits( $data->image, 'cover' );
 
 	echo '</div>';
 }