+++ modules/notice/script.js
@@ -1,11 +1,10 @@
-'use strict';
-
 var Tela = window.Tela || {};
+Tela.modules = Tela.modules || {};
 
-Tela.notice = (function(){
+Tela.modules.notice = (function(){
 
   function module(selector){
-    var $el     = $(selector),
+    var $el = $(selector),
       $closeButton;
 
     function init(){
@@ -33,5 +32,5 @@ Tela.notice = (function(){
 })();
 
 $(document).ready(function(){
-  Tela.notice('.notice');
+  Tela.modules.notice('.notice');
 });