+++ modules/search-box/search-box.php
@@ -1,8 +1,9 @@
 <?php function telabotanica_module_search_box($data) {
 	$defaults = [
 		'autocomplete' => true,
+		'instantsearch' => false,
 		'placeholder' => __('Rechercher une plante, un projet, un mot clé...', 'telabotanica'),
-		'value' => get_search_query(),
+		'value' => get_search_query() ?: get_query_var( 'q', false ),
 		'index' => false,
 		'suggestions' => false,
 		'modifiers' => ['large']
@@ -12,9 +13,11 @@
 	$data->modifiers = telabotanica_styleguide_modifiers_array('search-box', $data->modifiers);
 
 	printf(
-		'<div class="%s" data-autocomplete="%s">',
+		'<div class="%s" data-autocomplete="%s" data-instantsearch="%s" data-index="%s">',
 		implode(' ', $data->modifiers),
-		var_export($data->autocomplete, true)
+		var_export($data->autocomplete, true),
+		var_export($data->instantsearch, true),
+		$data->index
 	);
 		printf(
 			'<form role="search" method="get" action="%s" class="search-box-wrapper">',
@@ -22,13 +25,13 @@
 		);
 			printf(
 				'<input name="s" type="text" class="search-box-input" placeholder="%s" value="%s" autocomplete="off" spellcheck="false" />',
-				$data->placeholder,
-				$data->value
+				esc_attr( $data->placeholder ),
+				esc_attr( $data->value )
 			);
 			if ($data->index) :
 				printf(
-					'<input name="index" type="hidden" value="%s" />',
-					$data->index
+					'<input name="in" type="hidden" value="%s" />',
+					esc_attr( sanitize_key( $data->index ) )
 				);
 			endif;
 			printf(