+++ modules/share/script.js
@@ -1,10 +1,9 @@
-'use strict';
-
 var Tela = window.Tela || {};
+Tela.modules = Tela.modules || {};
 
-Tela.share = (function(){
+Tela.modules.share = (function(){
 
-  function share(selector){
+  function module(selector){
     var $el     = $(selector),
       $links;
 
@@ -15,6 +14,7 @@ Tela.share = (function(){
     }
 
     function onClickLink(e){
+      if ($(this).attr('target') != '_blank') {return;}
       e.preventDefault();
       e.stopPropagation();
       openInNewWindow($(this).attr('href'));
@@ -32,12 +32,12 @@ Tela.share = (function(){
 
   return function(selector){
     return $(selector).each(function(){
-      share(this);
+      module(this);
     });
   };
 
 })();
 
 $(document).ready(function(){
-  Tela.share('.share');
+  Tela.modules.share('.share');
 });